client-webapp/pyproject.toml
[project]
name = "client-webapp"
version = "0.1.0"
description = ""
authors = [
    {name = "ianba",email = "ianbabington93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.126.0,<0.127.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
black = "^25.12.0"
ruff = "^0.14.14"
pyright = "^1.1.408"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
